Refactor project structure into include/dotmake
