+++ include/dotmake/link_task.cpp
@@ -1,4 +1,4 @@
-#include "link_task.h"
+#include "dotmake/link_task.h"
 
 dotmake::LinkTask::LinkTask(const LinkTask& link_task,
                             AbstractSystemCaller& abstract_system_caller)
