+++ include/dotmake/shell_task_parser.h
@@ -12,9 +12,9 @@
 #include <string>
 #include <vector>
 
-#include "abstract_yaml_parser.h"
-#include "shell_task.h"
-#include "task.h"
+#include "dotmake/abstract_yaml_parser.h"
+#include "dotmake/shell_task.h"
+#include "dotmake/task.h"
 #include "yaml-cpp/node/node.h"
 #include "yaml-cpp/node/parse.h"
 #include "yaml-cpp/yaml.h"
