+++ test/link_task_parser.cpp
@@ -1,10 +1,10 @@
+#include <dotmake/link_task_parser.h>
 #include <gtest/gtest.h>
-#include <link_task_parser.h>
 
 #include <filesystem>
 
+#include "dotmake/system_caller.h"
 #include "mock_system_caller.h"
-#include "system_caller.h"
 
 using ::testing::Exactly;
 using ::testing::Return;
