+++ test/mock_system_caller.h
@@ -3,7 +3,7 @@
 
 #include <gmock/gmock.h>
 
-#include "abstract_system_caller.h"
+#include "dotmake/abstract_system_caller.h"
 
 class MockSystemCaller : public AbstractSystemCaller {
  public:
