+++ test/shell_task_parser.cpp
@@ -1,5 +1,5 @@
+#include <dotmake/shell_task_parser.h>
 #include <gtest/gtest.h>
-#include <shell_task_parser.h>
 
 #include "mock_system_caller.h"
